backend/chatbot/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP client
httpx==0.28.1

# LLM
openai==1.55.3

# GCP (optional - for GCS storage, Pub/Sub)
google-cloud-storage==2.18.2
google-cloud-pubsub==2.21.0

# Utils
python-dotenv==1.0.1
